# REF: Optimize patch adjoint operation