--- a/src/tike/operators/cupy/convolution.cu
+++ b/src/tike/operators/cupy/convolution.cu
@@ -50,6 +50,9 @@ _adjoint(float2 *patches, float2 *images, int nimagex, int pi, int ii,
 // patch pixels is much slower. There are many more checks and conditionals in
 // order to avoid atomic operations.
 
+// NOTE: nscan is the first grid dimension because it is has a larger maximum
+// (2^31) than the later two (6k).
+
 // The kernel should be launched with the following maximum shapes:
 // grid shape = (nscan, nimage, patch_size)
 // block shape = (min(max_thread, patch_size), 1, 1)