<!--Thank you for opening a pull request!

We appreciate that you care about this project enough to fix it, and we welcome contributions. We will make every effort to review your pull request in a timely manner. Please help us by following the instructions below.
-->

## Purpose
<!--Describe the problem or feature.

Only address one feature per pull request. If the scope of a single pull request is small (less than 400 lines of code), reviewers can understand it more quickly.

Link to any existing Issues. Use the `Closes` keyword if this Pull closes any Issues.
-->
Profiled some reconstructions and the patch adjoint kernel was the most time function. I tried to make that kernel faster, but I didn't really make any significant differences. This is just a result of some minor tweaks that I made along the way.

The main outcome of this tweaking was that I added an extra parameter to the kernel that allows for a new kind of broadcasting in order to avoid a memory copy in certain places in the code.

## Approach
<!--Describe how your changes address the problem.

Provide a brief summary of your algorithm(s) here.
-->
I used the CuPy wrapper API for NVProf to generate some timing reports:

```
Type  Time(%)      Time     Calls       Avg       Min       Max  Name
 GPU activities:   29.26%  1.05751s      1600  660.94us  624.57us  1.0273ms  adj_patch
                   14.53%  525.20ms      4032  130.26us  1.2800us  760.16us  cupy_multiply__complex64_complex64_complex64
                    6.67%  240.99ms      1152  209.19us  12.160us  1.4604ms  cupy_sum
                    5.98%  216.17ms       320  675.54us  671.29us  692.73us  void vector_fft<unsigned int=128, unsigned int=8, unsigned int=2, unsigned int=32, padding_t=6, twiddle_t=0, loadstore_modifier_t=2, layout_t=0, unsigned int, float>(kernel_arguments_t<unsigned int>)
                    5.95%  215.02ms       320  671.93us  667.45us  852.80us  cupy_true_divide__complex64_complex_complex64

```
The most time was spent on the adjoint patch kernel, so I tried to make it more efficient. When that didn't work, I looked into reducing the number of function calls. This lead me to adjusting the API to allow avoiding calling cp.tile in a few places. This should reduce memory overhead. In a separate PR, I should look into reducing the number of calls to multiplications and sums.

## Pre-Merge Checklists

### Submitter
- [x] Write a helpfully descriptive pull request title.
- [x] Organize changes into logically grouped commits with descriptive commit messages.
- [x] Document all new functions.
- [ ] Click 'details' on the readthedocs check to view the updated docs.
- [x] Write tests for new functions or explain why they are not needed.
- [ ] Address any complaints from pep8speaks.

### Reviewer
- [ ] Actually read all of the code.
- [ ] Run the new code yourself; the included tests should make this easy.
- [ ] Write a summary of the changes as you understand them.
- [ ] Thank the submitter.